Added flag to ensure Solr isn't reinstalled everytime vagrant is reprovisioned
Issue #9

+++ bootstrap.sh
@@ -5,10 +5,14 @@
 ## Please keep in alphabetical order.
 ##
 
+# The build of Solr to obtain
 SOLR_BUILD=http://archive.apache.org/dist/lucene/solr/4.3.1/solr-4.3.1.tgz
 
 # Bring apt-get up to date
 apt-get update
+
+# Install Java
+apt-get install -y openjdk-7-jdk
 
 # Git
 apt-get install -y git
@@ -19,19 +23,33 @@ apt-get install -y vim
 # Create source directory
 mkdir -p "/vagrant/source"
 
-# Own as user
+# Own source directory as user
 chown -R vagrant:vagrant "/vagrant/source/"
 
-# Link source directory into home 
+# Link persistent source directory into home 
 if [[ ! -L "/home/vagrant/source" ]]; then
   ln -s "/vagrant/source" "/home/vagrant/source"
 fi
 
-# Download Solr
-wget ${SOLR_BUILD}
-tar -zxvf solr-4.3.1.tgz
-rm solr-4.3.1.tgz
-cd solr-4.3.1
-apt-get install -y openjdk-7-jdk
-cd example
+# Download Solr and install only if it's the first time provisioning
+# indicated by whether or not .solrinstalled is present in / directory
+if [ ! -f /.solrinstalled ]; then
+  # Download
+  wget ${SOLR_BUILD}
+  # Unzip
+  tar -zxvf solr-4.3.1.tgz
+  # Remove the compressed file
+  rm solr-4.3.1.tgz
+  # Write file to indicate Solr has been installed
+  touch /.solrinstalled
+
+  echo "Provisioning Solr."
+else
+  # Nothing needs to be done if /.solrinstalled is present
+  echo "Solr already installed."
+fi
+
+# Enter the Jetty jar server directory
+cd solr-4.3.1/example
+# Run Solr using the Jetty server
 java -jar start.jar &
